+++ biblioteca/LuzAmbiente.cpp
@@ -1,7 +1,7 @@
 
 #include "LuzAmbiente.hpp"
 
-LuzAmbiente::LuzAmbiente(int r, int g, int b, int x, int y, int z): Luz(r, g, b){
+LuzAmbiente::LuzAmbiente(float r, float g, float b, double x, double y, double z): Luz(r, g, b){
     posicaoAmbiente = new Ponto();
     posicaoAmbiente->x = x;
     posicaoAmbiente->y = y;
